# openapi.yaml
openapi: 3.1.0
info:
  title: Paklog Inventory Management API
  description: |
    # Overview
    Comprehensive API for managing and querying inventory levels across the fulfillment network.

    ## Core Features
    - **Real-time Stock Queries**: Get current stock levels with caching for performance
    - **Stock Adjustments**: Manual adjustments with audit trail
    - **Bulk Operations**: High-performance batch processing (1,000-10,000+ items)
    - **Multi-Location Support**: Warehouse and location-based inventory tracking
    - **Stock Transfers**: Inter-location transfer management
    - **Stock Status Management**: Quarantine, damage, expiry tracking
    - **Serial Number Tracking**: Individual item tracking
    - **Inventory Valuation**: Cost tracking and COGS calculations
    - **Inventory Health**: Turnover analysis and dead stock detection
    - **CloudEvents Integration**: Event-driven architecture with standardized events

    ## Performance
    - L1 Cache (Caffeine): Sub-millisecond response times
    - L2 Cache (Redis): Distributed caching across instances
    - Async Processing: Non-blocking operations for metrics and events
    - MongoDB Aggregation: Server-side processing for complex queries

    ## Observability
    - OpenTelemetry distributed tracing
    - Prometheus metrics exposition
    - Structured logging with correlation IDs

    ## API Groups
    - **Core Inventory**: Stock levels, adjustments, allocations
...
    - **Holds**: Inventory holds and blocks
...
servers:
...
paths:
...
  /inventory/stock_levels/{sku}:
    get:
      summary: Get Stock Level for a SKU
      description: |
        Retrieves the current inventory levels for a specific SKU.
        Returns quantity on hand, allocated quantity, and available to promise.
      operationId: get_stock_level_by_sku
      tags: [Core Inventory]
      parameters:
        - name: sku
          in: path
          required: true
          description: The unique seller SKU to query
          schema:
            type: string
          examples:
            standard:
              value: "SKU-12345"
            withDashes:
              value: "PROD-ABC-001"
      responses:
        '200':
          description: OK. The current stock level is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockLevel'
              examples:
                inStock:
                  summary: Product in stock
                  value:
                    sku: "SKU-12345"
                    quantity_on_hand: 500
                    quantity_allocated: 150
                    available_to_promise: 350
                    location_id: "WH-001"
                lowStock:
                  summary: Low stock product
                  value:
                    sku: "SKU-67890"
                    quantity_on_hand: 10
                    quantity_allocated: 8
                    available_to_promise: 2
                    location_id: "WH-002"
                outOfStock:
                  summary: Out of stock product
                  value:
                    sku: "SKU-99999"
                    quantity_on_hand: 0
                    quantity_allocated: 0
                    available_to_promise: 0
                    location_id: "WH-001"
        '404':
          description: Not Found. No stock record found for the given SKU.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Not Found"
                message: "No stock record found for SKU: SKU-INVALID"
                timestamp: "2025-11-09T20:00:00Z"
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Adjust Stock Level (Relative)
      description: |
        Allows manual adjustment to the quantity on hand for a SKU using relative changes.
        This is the industry-standard PATCH operation for incremental adjustments.
      operationId: adjust_stock_level
      tags: [Core Inventory]
      parameters:
        - name: sku
          in: path
          required: true
          description: The unique seller SKU to adjust
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockLevelRequest'
            examples:
              increaseStock:
                summary: Increase stock
                value:
                  quantity_change: 100
                  reason_code: "PURCHASE_RECEIPT"
                  comment: "PO-12345 received"
              decreaseStock:
                summary: Decrease stock
                value:
                  quantity_change: -25
                  reason_code: "DAMAGE"
                  comment: "Water damage in aisle 3"
      responses:
        '200':
          description: OK. The adjustment was successful and the updated stock level is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockLevel'
        '400':
          description: Bad Request. The request body is invalid or business rules violated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found. No stock record found for the SKU to be adjusted.
        '500':
          description: Internal Server Error.
  /inventory/stock_levels/{sku}/set:
    put:
      summary: Set Absolute Stock Level (Physical Count)
      description: |
        Sets the exact stock quantity for a SKU, typically used for physical inventory counts.
        Industry standard pattern from Shopify, WooCommerce, SAP, and Oracle WMS.

        **Use Cases:**
        - Physical inventory counts
        - Cycle count corrections
        - Initial stock setup
        - System corrections from WMS/ERP integration

        **Difference from PATCH /stock_levels/{sku}:**
        - PATCH: Relative adjustment (add/subtract quantity)
        - PUT: Absolute set (override to exact quantity)
      operationId: set_stock_level
      tags: [Core Inventory]
      parameters:
        - name: sku
          in: path
          required: true
          description: The unique seller SKU
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetStockLevelRequest'
            examples:
              physicalCount:
                summary: Physical count result
                value:
                  quantity: 245
                  reason_code: "PHYSICAL_COUNT"
                  comment: "Annual inventory count 2025"
                  location_id: "WH-001"
                  count_date: "2025-11-09T10:00:00Z"
              cycleCount:
                summary: Cycle count adjustment
                value:
                  quantity: 98
                  reason_code: "CYCLE_COUNT"
                  comment: "Weekly cycle count - Zone A"
                  location_id: "WH-EAST"
              wmsSync:
                summary: WMS system sync
                value:
                  quantity: 1500
                  reason_code: "INTEGRATION_SYNC"
                  source_system: "WMS"
                  source_transaction_id: "WMS-TXN-789456"
                  location_id: "DC-CENTRAL"
      responses:
        '200':
          description: OK. Stock level successfully set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockLevel'
              example:
                sku: "SKU-12345"
                quantity_on_hand: 245
                quantity_allocated: 50
                available_to_promise: 195
                location_id: "WH-001"
        '400':
          description: Bad Request. Invalid quantity or business rule violation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found. SKU not found.
        '500':
          description: Internal Server Error.
  /inventory/stock_levels/batch:
    post:
      summary: Batch Stock Updates
      description: |
        Process multiple stock updates in a single request.
        Industry standard pattern for high-volume operations.

        **Capabilities:**
        - Up to 1,000 updates per batch
        - Mix of ADJUST (relative) and SET (absolute) operations
        - Partial success handling (one failure doesn't stop others)
        - Source system tracking for each update

        **Performance:**
        - Optimized for bulk WMS/ERP integrations
        - Async processing for large batches
        - Detailed per-item results

        **Use Cases:**
        - WMS bulk synchronization
        - ERP integration batches
        - Physical count imports
        - Multi-warehouse updates
      operationId: batch_stock_update
      tags: [Bulk Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchStockUpdateRequest'
            examples:
              mixedBatch:
                summary: Mixed adjustments and sets
                value:
                  batch_id: "BATCH-2025-11-09-001"
                  source_system: "WMS"
                  updates:
                    - sku: "SKU-001"
                      update_type: "ADJUST"
                      quantity: 50
                      reason_code: "PURCHASE_RECEIPT"
                      comment: "PO-12345 received"
                    - sku: "SKU-002"
                      update_type: "SET"
                      quantity: 100
                      reason_code: "PHYSICAL_COUNT"
                      location_id: "WH-001"
                    - sku: "SKU-003"
                      update_type: "ADJUST"
                      quantity: -10
                      reason_code: "DAMAGE"
              wmsIntegration:
                summary: WMS bulk sync
                value:
                  batch_id: "WMS-SYNC-20251109-0900"
                  source_system: "WMS"
                  updates:
                    - sku: "SKU-101"
                      update_type: "SET"
                      quantity: 450
                      reason_code: "INTEGRATION_SYNC"
                      source_transaction_id: "WMS-101"
                      location_id: "DC-EAST"
                    - sku: "SKU-102"
                      update_type: "SET"
                      quantity: 320
                      reason_code: "INTEGRATION_SYNC"
                      source_transaction_id: "WMS-102"
                      location_id: "DC-EAST"
      responses:
        '200':
          description: OK. Batch processing completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStockUpdateResponse'
              example:
                batch_id: "BATCH-2025-11-09-001"
                total_requested: 3
                successful_updates: 2
                failed_updates: 1
                processing_time_ms: 156
                results:
                  - sku: "SKU-001"
                    success: true
                    previous_quantity: 100
                    new_quantity: 150
                  - sku: "SKU-002"
                    success: true
                    previous_quantity: 95
                    new_quantity: 100
                  - sku: "SKU-003"
                    success: false
                    error_message: "Insufficient stock for adjustment"
        '400':
          description: Bad Request. Invalid batch format or size exceeds limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
  /inventory/stock_levels/{sku}/locations/{locationId}:
    get:
      summary: Get Stock Level by Location
      description: |
        Retrieves stock level for a specific SKU at a specific location.
        Multi-location support following industry patterns from Shopify, SAP, Oracle.
      operationId: get_stock_level_by_location
      tags: [Core Inventory]
      parameters:
        - name: sku
          in: path
          required: true
          description: The unique seller SKU
          schema:
            type: string
        - name: locationId
          in: path
          required: true
          description: Location/warehouse identifier
          schema:
            type: string
          examples:
            warehouse:
              value: "WH-001"
            store:
              value: "STORE-NYC"
            dc:
              value: "DC-CENTRAL"
      responses:
        '200':
          description: OK. Stock level at location returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockLevel'
              example:
                sku: "SKU-12345"
                quantity_on_hand: 500
                quantity_allocated: 150
                available_to_promise: 350
                location_id: "WH-001"
        '404':
          description: Not Found. SKU or location not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
  /inventory/adjustments:
    post:
      summary: Create Stock Adjustment
      description: |
        Creates a stock adjustment record.
        Alternative endpoint to PATCH /stock_levels/{sku} with explicit adjustment tracking.
      operationId: create_stock_adjustment
      tags: [Core Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdjustmentRequest'
            example:
              sku: "SKU-12345"
              quantity_change: -15
              reason_code: "DAMAGE"
              comment: "Damaged during picking"
      responses:
        '202':
          description: Accepted. Stock adjustment is being processed.
        '400':
          description: Bad Request. Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found. SKU does not exist.
        '500':
          description: Internal Server Error.
  /inventory/stock_levels/{sku}/reservations:
    post:
      summary: Create Stock Reservation
      description: |
        Reserves (allocates) stock for a specific order.
        This reduces the available-to-promise quantity without physically picking the item.
      operationId: create_reservation
      tags: [Core Inventory]
      parameters:
        - name: sku
          in: path
          required: true
          description: The unique seller SKU to reserve stock for
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
            example:
              quantity: 5
              order_id: "ORD-12345"
      responses:
        '202':
          description: Accepted. Stock reservation is being processed.
        '400':
          description: Bad Request. Invalid request or insufficient stock.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found. SKU does not exist.
        '500':
          description: Internal Server Error.
  /inventory/allocations/bulk:
    post:
      summary: Bulk Allocate Stock
      description: |
        High-performance bulk allocation endpoint for processing large order batches.
        Optimized to handle 10,000+ allocation requests in a single API call.

        **Performance Characteristics:**
        - Batch processing with parallel execution
        - Database connection pooling
        - Optimistic locking for concurrency
        - Partial success handling

        **Use Cases:**
        - Order wave processing
        - Batch order import
        - Marketplace bulk fulfillment
      operationId: bulk_allocate_stock
      tags: [Bulk Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAllocationRequest'
            example:
              requests:
                - sku: "SKU-12345"
                  quantity: 10
                  order_id: "ORD-001"
                - sku: "SKU-67890"
                  quantity: 5
                  order_id: "ORD-002"
                - sku: "SKU-11111"
                  quantity: 20
                  order_id: "ORD-003"
      responses:
        '200':
          description: OK. Bulk allocation completed with results for each item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAllocationResponse'
              example:
                successful_allocations: 2
                failed_allocations: 1
                processing_time_ms: 450
                results:
                  - sku: "SKU-12345"
                    order_id: "ORD-001"
                    success: true
                    allocated_quantity: 10
                  - sku: "SKU-67890"
                    order_id: "ORD-002"
                    success: true
                    allocated_quantity: 5
                  - sku: "SKU-11111"
                    order_id: "ORD-003"
                    success: false
                    error_message: "Insufficient stock: available=15, requested=20"
        '400':
          description: Bad Request. Invalid request format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
...
components:
  schemas:
...
    StockLevel:
      type: object
      description: Current stock level for a SKU
      properties:
        sku:
          type: string
          description: The unique identifier for the product
        quantity_on_hand:
          type: integer
          format: int32
          description: The total physical quantity in the warehouse
          minimum: 0
        quantity_allocated:
          type: integer
          format: int32
          description: Quantity reserved for orders not yet picked
          minimum: 0
        available_to_promise:
          type: integer
          format: int32
          description: Quantity available for new sales (on_hand - allocated)
          minimum: 0
        location_id:
          type: string
          description: Warehouse/location identifier (e.g., WH-001, STORE-NYC)
          nullable: true
      required:
        - sku
        - quantity_on_hand
        - quantity_allocated
        - available_to_promise
    UpdateStockLevelRequest:
      type: object
      description: Request to adjust stock level (relative change)
      required:
        - quantity_change
        - reason_code
      properties:
        quantity_change:
          type: integer
          format: int32
          description: Change to apply (positive for increase, negative for decrease)
        reason_code:
          type: string
          description: Reason for adjustment
          enum:
            - PURCHASE_RECEIPT
            - RETURN_TO_STOCK
            - TRANSFER_IN
            - PRODUCTION_COMPLETE
            - SALE
            - DAMAGE
            - THEFT_LOSS
            - TRANSFER_OUT
            - DISPOSAL
            - PHYSICAL_COUNT
            - CYCLE_COUNT
            - SYSTEM_CORRECTION
            - ITEM_PICKED
            - ALLOCATION
            - DEALLOCATION
            - STOCK_RECEIPT
        comment:
          type: string
          description: Optional comment explaining the adjustment
    SetStockLevelRequest:
      type: object
      description: Request to set absolute stock quantity (physical count)
      required:
        - quantity
        - reason_code
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 0
          description: Exact stock quantity to set
        reason_code:
          type: string
          description: Reason for setting absolute quantity
          enum:
            - PHYSICAL_COUNT
            - CYCLE_COUNT
            - SYSTEM_CORRECTION
            - INTEGRATION_SYNC
            - INITIAL_STOCK
        comment:
          type: string
          description: Optional comment
        location_id:
          type: string
          description: Warehouse/location identifier
        count_date:
          type: string
          format: date-time
          description: When the physical count was performed
        source_system:
          type: string
          description: Source system name (e.g., WMS, ERP)
        source_transaction_id:
          type: string
          description: Transaction ID in source system
        source_operator_id:
          type: string
          description: Operator who performed the count
    BatchStockUpdateRequest:
      type: object
      description: Request for batch stock updates
      required:
        - updates
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/StockUpdateItem'
          minItems: 1
          maxItems: 1000
          description: List of stock updates (max 1000)
        batch_id:
          type: string
          description: Optional batch tracking ID
        source_system:
          type: string
          description: Source system for all updates
    StockUpdateItem:
      type: object
      description: Individual stock update within a batch
      required:
        - sku
        - update_type
        - quantity
        - reason_code
      properties:
        sku:
          type: string
        update_type:
          type: string
          enum: [ADJUST, SET]
          description: ADJUST for relative, SET for absolute
        quantity:
          type: integer
          format: int32
        reason_code:
          type: string
        comment:
          type: string
        location_id:
          type: string
        source_transaction_id:
          type: string
    BatchStockUpdateResponse:
      type: object
      description: Response from batch stock update
      properties:
        batch_id:
          type: string
        total_requested:
          type: integer
        successful_updates:
          type: integer
        failed_updates:
          type: integer
        processing_time_ms:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/UpdateResult'
    UpdateResult:
      type: object
      description: Result for individual update in batch
      properties:
        sku:
          type: string
        success:
          type: boolean
        error_message:
          type: string
        previous_quantity:
          type: integer
        new_quantity:
          type: integer
    CreateAdjustmentRequest:
      type: object
      description: Request to create stock adjustment
      required:
        - sku
        - quantity_change
        - reason_code
      properties:
        sku:
          type: string
        quantity_change:
          type: integer
        reason_code:
          type: string
        comment:
          type: string
    CreateReservationRequest:
      type: object
      description: Request to create stock reservation/allocation
      required:
        - quantity
        - order_id
      properties:
        quantity:
          type: integer
          minimum: 1
        order_id:
          type: string
    BulkAllocationRequest:
      type: object
      description: Bulk allocation request (up to 10,000 items)
      required:
        - requests
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/AllocationRequestItem'
          minItems: 1
          maxItems: 10000
    AllocationRequestItem:
      type: object
      required:
        - sku
        - quantity
        - order_id
      properties:
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        order_id:
          type: string
    BulkAllocationResponse:
      type: object
      description: Bulk allocation response
      properties:
        successful_allocations:
          type: integer
        failed_allocations:
          type: integer
        processing_time_ms:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/AllocationResult'
    AllocationResult:
      type: object
      properties:
        sku:
          type: string
        order_id:
          type: string
        success:
          type: boolean
        allocated_quantity:
          type: integer
        error_message:
          type: string
...
    Error:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: When error occurred
        traceId:
          type: string
          description: Distributed trace ID for debugging